app/Services/QeWorker/requirements.txt
pika==1.3.2
unbabel-comet==2.2.2
torch